refactor(agent)!: rig-core knows nothing about agents — the run layer is rig_agent::run
Applying the strict test — rig-core is the provider, tool, transport and
message contract; nothing in it should know what an agent is — moves the
rest of the 'vocabulary' out of rig-core into rig_agent::run:

- run::spec (RunSpec), run::prepare (prepare_request, PreparedRequest),
  run::output (OutputMode + output-policy helpers, augment_preamble),
  run::patch (RequestPatch): how an agent turns its spec into a request and
  its structured-output strategy — agent behaviour, not provider contract;
- run::transcript: history threading, invalid-call feedback text, turn
  classification — loop decisions.

rig-core keeps exactly two things, because they are properties of the
message model: transcript::validate_canonical (+ TranscriptError) and the
tool-result constructors tool_result_output / tool_result_message.

Consequences: the rig-core-only 'vocabulary driver' fixture loses its point
and is removed; tests/fixtures/agent_run_stepper (rig-agent with default
features off + rig-core, guarded runtime-free) is the second-driver
evidence, and with exactly one prepare_request every driver sends the same
bytes by construction. rig::run is unchanged (rig_agent::run behind agent).
The no-async guard follows the files; the 0.42 path pin is unchanged.
