+++ CHANGELOG.md
@@ -14,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - *(core)* owned-future entry points for hosts that spawn tasks: `ToolCatalog::execute_owned` (returns `(ToolResult, ToolContext)`) and `ConversationMemoryExt::{load_owned, append_owned, clear_owned}` over `Arc<Self>` + owned `ConversationId`
 - *(core)* [**breaking**] `ModelHandle` moves to `rig_core::completion` (re-exported at every old rig-agent path) beside `EmbeddingModelHandle`; new `ModelRef` (`Arc<str>`, transparent serde) is the string identity a spec/registry names a model by — `ModelHandle::named` takes `impl Into<ModelRef>`, `model_ref()` reads it. See `MIGRATING.md`
 - *(core)* [**breaking**] the contextual tool API — `Tool`, `ToolEmbedding`, `ErasedTool`, `DynamicTool`, `RegisteredTool`, `ToolDispatch`, `dispatch_tool`, `tool_definition`, `ToolSet` — moves to `rig_core::tool` (re-exported at `rig_agent::tool::*` and at `rig::tool::*` without the `agent` feature); new `ToolCatalog` (definitions + dispatch by name, `ToolSet::catalog()`), of which rig-agent's `ToolRegistrySnapshot` is now an alias; `#[rig_tool]` expands contextual tools against rig-core, so a rig-core-only crate can author them. `ToolServer`/`ToolServerHandle` stay in rig-agent. See `MIGRATING.md`
-- *(core)* [**breaking**] `rig_core::completion::prepare::prepare_request` — the pure `(RunSpec, capabilities, history, tools, committed output tool, RequestPatch) -> PreparedRequest` step (output-mode resolution, output-tool synthesis, preamble augmentation, `active_tools`, tool-choice validation) with `PreparedRequest::apply(builder)`; the output-policy helpers it uses are public in `rig_core::completion::output`; `RequestPatch` moves to `rig_core::completion::patch` (re-exported at the old hook path). rig-agent's driver is rewired over it with byte-identical requests (golden test + recorded provider suites); `AgentRun::advertised_tools()` now carries the definitions the request actually sent (it was always empty). A guard runs a rig-core-only fixture driver end to end (`tests/core/core_run_driver.rs`). See `MIGRATING.md`
-- *(core, agent)* [**breaking**] the run protocol's vocabulary is rig-core's and `AgentRun` is `rig_agent::run`: `rig_core::transcript` (transcript helpers, `validate_canonical`) and `rig_core::completion::{output, patch, prepare, spec}` (`OutputMode`, `RequestPatch`, `prepare_request`, `RunSpec`) need no runtime and no `agent` feature; `rig_agent::run` holds `AgentRun`, `AgentRunStep`, `ModelTurn`, `PendingToolCall`, `RunEntry`, `TurnTools` and the run's own types — `PromptResponse`/`CompletionCall`/`PromptError` (`run::response`), the invalid-call/retry decisions (`run::policy`), the streamed-turn assembler (`run::streamed`) — still sans-IO (source-level guard) and serializable; `rig::run` is that module behind `agent`; every 0.42 rig-agent path resolves unchanged. Adds `RunSpec` (protocol config as data, `AgentRun::from_spec`, `Agent::run_spec`, `AgentBuilder::apply_spec`, `RunSpec::DEFAULT_OUTPUT_RETRIES`), `AgentRun::advertise_tools`/`advertised_tools` (`TurnTools`), `validate_canonical`/`AgentRun::with_validated_history`, and a `NonZeroU64` counter `RunId` in `rig_core::id` (`as_str` gone). See `MIGRATING.md`
+- *(agent)* [**breaking**] `rig_agent::run::prepare::prepare_request` — the pure `(RunSpec, capabilities, history, tools, committed output tool, RequestPatch) -> PreparedRequest` step (output-mode resolution, output-tool synthesis, preamble augmentation, `active_tools`, tool-choice validation) with `PreparedRequest::apply(builder)`; the output-policy helpers it uses are public in `rig_agent::run::output`; `RequestPatch` moves to `rig_agent::run::patch` (re-exported at the old hook path). rig-agent's driver is rewired over it with byte-identical requests (golden test + recorded provider suites); `AgentRun::advertised_tools()` now carries the definitions the request actually sent (it was always empty). A guard runs a fixture that steps `AgentRun` with rig-agent (default features off) + rig-core end to end and checks that graph is runtime-free (`tests/core/agent_run_stepper.rs`). See `MIGRATING.md`
+- *(core, agent)* [**breaking**] the run protocol is rig-agent's sans-IO `run` layer and rig-core knows nothing about agents: `rig_agent::run` holds `AgentRun`, `AgentRunStep`, `ModelTurn`, `PendingToolCall`, `RunEntry`, `TurnTools`, `RunSpec` (`run::spec`), `prepare_request` (`run::prepare`), `OutputMode` + output policy (`run::output`), `RequestPatch` (`run::patch`), `PromptResponse`/`CompletionCall`/`PromptError` (`run::response`), the invalid-call/retry decisions (`run::policy`), the streamed-turn assembler (`run::streamed`) and the loop-side transcript helpers (`run::transcript`) — all sans-IO (source-level guard) and serializable; rig-core keeps only the message-model invariants `rig_core::transcript::{validate_canonical, tool_result_output, tool_result_message}`; `rig::run` is that module behind `agent`; a second driver steps `AgentRun` via `rig-agent` with default features off, whose graph is guarded runtime-free; every 0.42 rig-agent path resolves unchanged. Adds `RunSpec` (protocol config as data, `AgentRun::from_spec`, `Agent::run_spec`, `AgentBuilder::apply_spec`, `RunSpec::DEFAULT_OUTPUT_RETRIES`), `AgentRun::advertise_tools`/`advertised_tools` (`TurnTools`), `validate_canonical`/`AgentRun::with_validated_history`, and a `NonZeroU64` counter `RunId` in `rig_core::id` (`as_str` gone). See `MIGRATING.md`
 - *(core)* `http_client::BoxedHttpClient`, a type-erased `HttpClientExt` (`Arc<dyn …>`, cheap `Clone`, byte-transparent) so a host can hold one transport for every provider without naming it; `Client<Ext, H>` now defaults `H` to it in type position, `ProviderFromEnv::{from_env_boxed, from_val_boxed}`, `Client::boxed`, and `rig_reqwest::ReqwestClient::boxed()`. See `MIGRATING.md` for the one break (`Client::<Ext>::builder()` → `Client::<Ext, Missing>::builder()`)
 - *(llamacpp)* [**breaking**] `providers::llamacpp` replaces `providers::llamafile`, named after the upstream everyone runs rather than Mozilla's single-file distribution — the same server, so a `.llamafile` is reached by pointing `llamacpp::Client` at it. The old module is deleted, not aliased; see `MIGRATING.md`. The new provider adds an optional API key (so a `llama-server --api-key` deployment is reachable at all), `model_listing` and `rerank` capability slots, a `build_uri` that no longer produces `/v1/v1`, and a `llamacpp::CompletionResponse` carrying llama.cpp's `timings` (by [gold-silver-copper](https://github.com/gold-silver-copper))
 - *(core)* a shared Jina-shaped rerank driver (`providers::internal::rerank`), so a provider on that wire declares a capability slot instead of hand-rolling a request builder; rig had `RerankModel` and exactly one implementation of it before this (by [gold-silver-copper](https://github.com/gold-silver-copper))
