+++ crates/rig-agent/src/run/output.rs
