+++ src/lib.rs
@@ -85,13 +85,11 @@ pub mod core {
     pub use rig_core::*;
 }
 
-/// The sans-IO run protocol: `AgentRun`, its step/turn types, the run's
-/// response and error types, the invalid-call decision data and the
-/// streamed-turn assembler (`rig_agent::run`). The request vocabulary a run
-/// is built from — `prepare_request`, `RunSpec`, `OutputMode`, `RequestPatch`,
-/// the transcript helpers — is rig-core's and needs no feature:
-/// `rig::core::completion::{prepare, spec, output, patch}`,
-/// `rig::core::transcript`.
+/// The sans-IO run layer of rig-agent (`rig_agent::run`): `AgentRun` and its
+/// step/turn types, the run's spec, request preparation, output policy and
+/// per-turn patch, its response and error types, the invalid-call decision
+/// data, the streamed-turn assembler and the loop-side transcript helpers.
+/// rig-core keeps only the message-model invariants (`rig::core::transcript`).
 #[cfg(feature = "agent")]
 #[cfg_attr(docsrs, doc(cfg(feature = "agent")))]
 pub mod run {
