+++ tests/core/mod.rs
@@ -1,5 +1,4 @@
 mod agent_run_stepper;
-mod core_run_driver;
 mod dependency_graph;
 #[cfg(feature = "derive")]
 mod embed_macro;
