--- tests/fixtures/core_run_driver/.gitignore
@@ -1,2 +1,0 @@
-/target
-Cargo.lock
