--- tests/fixtures/core_run_driver/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rig-core-run-driver-fixture"
-version = "0.0.0"
-edition = "2024"
-publish = false
-
-[workspace]
-
-# A hand-rolled driver over the run vocabulary with NO rig-agent and NO
-# AgentRun: rig-core alone (erased model, erased tool set, catalog, RunSpec,
-# prepare_request, transcript helpers). `tests/core/core_run_driver.rs` runs
-# this binary and checks its dependency graph; keep everything else out of it.
-[dependencies]
-rig-core = { path = "../../../crates/rig-core", default-features = false }
-serde_json = "1.0.150"
